+++ src/ScottPlot5/ScottPlot5/MultiplotLayouts/Columns.cs
@@ -2,12 +2,12 @@
 
 public class Columns : IMultiplotLayout
 {
-    public PixelRect[] GetSubplotRectangles(Multiplot multiplot, PixelRect figureRect)
+    public PixelRect[] GetSubplotRectangles(SubplotCollection subplots, PixelRect figureRect)
     {
-        PixelRect[] rectangles = new PixelRect[multiplot.Count];
+        PixelRect[] rectangles = new PixelRect[subplots.Count];
 
-        double fractionPerColumn = 1.0 / multiplot.Count;
-        for (int i = 0; i < multiplot.Count; i++)
+        double fractionPerColumn = 1.0 / subplots.Count;
+        for (int i = 0; i < subplots.Count; i++)
         {
             FractionRect fr = new(fractionPerColumn * i, 0, fractionPerColumn, 1);
             rectangles[i] = fr.GetPixelRect(figureRect);