+++ src/ScottPlot5/ScottPlot5/MultiplotLayouts/CustomGrid.cs
@@ -4,11 +4,11 @@ public class CustomGrid : IMultiplotLayout
 {
     readonly Dictionary<Plot, GridCell> PlotCells = [];
 
-    public PixelRect[] GetSubplotRectangles(Multiplot multiplot, PixelRect figureRect)
+    public PixelRect[] GetSubplotRectangles(SubplotCollection subplots, PixelRect figureRect)
     {
-        PixelRect[] rectangles = new PixelRect[multiplot.Count];
+        PixelRect[] rectangles = new PixelRect[subplots.Count];
 
-        Plot[] plots = multiplot.GetPlots();
+        Plot[] plots = subplots.GetPlots();
 
         for (int i = 0; i < plots.Length; i++)
         {