+++ src/ScottPlot5/ScottPlot5/MultiplotLayouts/Rows.cs
@@ -2,12 +2,12 @@
 
 public class Rows : IMultiplotLayout
 {
-    public PixelRect[] GetSubplotRectangles(Multiplot multiplot, PixelRect figureRect)
+    public PixelRect[] GetSubplotRectangles(SubplotCollection subplots, PixelRect figureRect)
     {
-        PixelRect[] rectangles = new PixelRect[multiplot.Count];
+        PixelRect[] rectangles = new PixelRect[subplots.Count];
 
-        double fractionPerRow = 1.0 / multiplot.Count;
-        for (int i = 0; i < multiplot.Count; i++)
+        double fractionPerRow = 1.0 / subplots.Count;
+        for (int i = 0; i < subplots.Count; i++)
         {
             FractionRect fr = new(0, fractionPerRow * i, 1, fractionPerRow);
             rectangles[i] = fr.GetPixelRect(figureRect);